Add files via upload
Screen 1:

This screen is used as a launch screen. In the UI we made a basic design. In UX, we implemented brand icon and background colour showing all sorts of workouts, nutrition and heart monitor, which shows how important sleep is for human body. Along with we added a label showing "Fitness Arena".